PaymentOptionsActvity & Viewmodel use FlowController Config
