+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentOptionsViewModelTest.kt
@@ -37,9 +37,10 @@ import com.stripe.android.model.PaymentMethodCreateParamsFixtures.DEFAULT_CARD
 import com.stripe.android.model.PaymentMethodFixtures
 import com.stripe.android.model.PaymentMethodFixtures.CARD_PAYMENT_SELECTION
 import com.stripe.android.model.PaymentMethodFixtures.toDisplayableSavedPaymentMethod
-import com.stripe.android.paymentsheet.PaymentSheetFixtures.updateState
 import com.stripe.android.paymentsheet.addresselement.AutocompleteContract
 import com.stripe.android.paymentsheet.analytics.EventReporter
+import com.stripe.android.paymentsheet.flowcontroller.FlowControllerFixtures
+import com.stripe.android.paymentsheet.flowcontroller.FlowControllerFixtures.updateState
 import com.stripe.android.paymentsheet.forms.FormFieldValues
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen
@@ -403,7 +404,7 @@ internal class PaymentOptionsViewModelTest {
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
-                    config = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
+                    config = FlowControllerFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
                         .paymentMethodLayout(PaymentMethodLayout.Vertical)
                         .build(),
                     paymentMethods = listOf(),
@@ -425,7 +426,7 @@ internal class PaymentOptionsViewModelTest {
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
-                    config = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
+                    config = FlowControllerFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
                         .paymentMethodLayout(PaymentMethodLayout.Vertical)
                         .build(),
                     paymentMethods = listOf(PaymentMethodFixtures.CARD_PAYMENT_METHOD),
@@ -447,7 +448,7 @@ internal class PaymentOptionsViewModelTest {
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
-                    config = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
+                    config = FlowControllerFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
                         .paymentMethodLayout(PaymentMethodLayout.Vertical)
                         .build(),
                     paymentMethods = listOf(),
@@ -469,7 +470,7 @@ internal class PaymentOptionsViewModelTest {
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
-                    config = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
+                    config = FlowControllerFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
                         .paymentMethodLayout(PaymentMethodLayout.Automatic)
                         .build(),
                     paymentMethods = listOf(),
@@ -491,7 +492,7 @@ internal class PaymentOptionsViewModelTest {
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
-                    config = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
+                    config = FlowControllerFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
                         .paymentMethodLayout(PaymentMethodLayout.Automatic)
                         .build(),
                     paymentMethods = listOf(PaymentMethodFixtures.CARD_PAYMENT_METHOD),
@@ -513,7 +514,7 @@ internal class PaymentOptionsViewModelTest {
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
-                    config = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
+                    config = FlowControllerFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
                         .paymentMethodLayout(PaymentMethodLayout.Automatic)
                         .build(),
                     paymentMethods = listOf(),
@@ -1205,7 +1206,7 @@ internal class PaymentOptionsViewModelTest {
                     isGooglePayReady = true,
                 ),
             ),
-            configuration = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY,
+            configuration = FlowControllerFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY,
             enableLogging = false,
             productUsage = mock(),
             paymentElementCallbackIdentifier = "PaymentOptionsViewModelTestCallbackIdentifier",
