+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/flowcontroller/FlowControllerFixtures.kt
@@ -1,10 +1,21 @@
 package com.stripe.android.paymentsheet.flowcontroller
 
+import com.stripe.android.common.model.asCommonConfiguration
 import com.stripe.android.elements.CustomerConfiguration
 import com.stripe.android.elements.payment.FlowController
 import com.stripe.android.elements.payment.PaymentMethodLayout
+import com.stripe.android.link.LinkAccountUpdate
+import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadataFactory
+import com.stripe.android.lpmfoundations.paymentmethod.WalletType
+import com.stripe.android.model.PaymentMethod
+import com.stripe.android.model.StripeIntent
 import com.stripe.android.paymentsheet.ConfigFixtures
+import com.stripe.android.paymentsheet.PaymentOptionContract
+import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.state.CustomerState
+import com.stripe.android.paymentsheet.state.LinkState
+import com.stripe.android.paymentsheet.state.PaymentSheetState
+import org.mockito.kotlin.mock
 
 internal object FlowControllerFixtures {
     internal const val MERCHANT_DISPLAY_NAME = "Merchant, Inc."
@@ -39,8 +50,60 @@ internal object FlowControllerFixtures {
         defaultPaymentMethodId = null,
     )
 
+    internal val CONFIG_GOOGLEPAY
+        get() = FlowController.Configuration(
+            merchantDisplayName = MERCHANT_DISPLAY_NAME,
+            googlePay = ConfigFixtures.GOOGLE_PAY,
+            paymentMethodLayout = PaymentMethodLayout.Horizontal,
+        )
+
     internal val CONFIG_CUSTOMER_WITH_GOOGLEPAY
         get() = CONFIG_CUSTOMER.newBuilder()
             .googlePay(ConfigFixtures.GOOGLE_PAY)
             .build()
+
+    internal val PAYMENT_OPTIONS_CONTRACT_ARGS = PaymentOptionContract.Args(
+        state = PaymentSheetState.Full(
+            customer = EMPTY_CUSTOMER_STATE,
+            config = CONFIG_GOOGLEPAY.asCommonConfiguration(),
+            paymentSelection = null,
+            validationError = null,
+            paymentMethodMetadata = PaymentMethodMetadataFactory.create(),
+        ),
+        configuration = CONFIG_GOOGLEPAY,
+        enableLogging = false,
+        productUsage = mock(),
+        paymentElementCallbackIdentifier = FLOW_CONTROLLER_CALLBACK_TEST_IDENTIFIER,
+        linkAccountInfo = LinkAccountUpdate.Value(null),
+        walletsToShow = WalletType.entries,
+    )
+
+    internal fun PaymentOptionContract.Args.updateState(
+        paymentMethods: List<PaymentMethod> = state.customer?.paymentMethods ?: emptyList(),
+        isGooglePayReady: Boolean = state.paymentMethodMetadata.isGooglePayReady,
+        stripeIntent: StripeIntent = state.stripeIntent,
+        config: FlowController.Configuration = configuration,
+        paymentSelection: PaymentSelection? = state.paymentSelection,
+        linkState: LinkState? = state.paymentMethodMetadata.linkState,
+    ): PaymentOptionContract.Args {
+        return copy(
+            state = state.copy(
+                customer = CustomerState(
+                    id = config.customer?.id ?: "cus_1",
+                    ephemeralKeySecret = config.customer?.ephemeralKeySecret ?: "client_secret",
+                    customerSessionClientSecret = null,
+                    paymentMethods = paymentMethods,
+                    defaultPaymentMethodId = null,
+                ),
+                config = config.asCommonConfiguration(),
+                paymentSelection = paymentSelection,
+                paymentMethodMetadata = PaymentMethodMetadataFactory.create(
+                    stripeIntent = stripeIntent,
+                    isGooglePayReady = isGooglePayReady,
+                    linkState = linkState,
+                ),
+            ),
+            configuration = config,
+        )
+    }
 }
